requirements.txt
pandas
numpy
scikit-learn
spotipy
python-dotenv
tqdm